+++ CorePayments/src/main/java/com/paypal/android/corepayments/browserswitch/BrowserSwitchLaunchMode.kt
@@ -1,0 +1,9 @@
+package com.paypal.android.corepayments.browserswitch
+
+import androidx.annotation.RestrictTo
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+enum class BrowserSwitchLaunchMode {
+    CUSTOM_TAB,
+    AUTH_TAB,
+}